Auth 파일 추가 및 개별 파일 삭제
